+++ packages/hooks_riverpod/lib/all.dart
@@ -1,0 +1,9 @@
+/// Exports all the implementation classes of providers, such as `Provider` vs `AutoDisposeProvider`.
+///
+/// This is useful if you want to use the `always_specify_types` lint. Otherwise,
+/// import `package:hooks_riverpod/riverpod.dart` instead, to avoid polluting the auto-complete.
+library all;
+
+export 'package:riverpod/all.dart';
+
+export 'hooks_riverpod.dart';