+++ packages/flutter_riverpod/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.10.1
+
+- Fixed invalid version error
+
 # 0.10.0
 
 - Fixed a bug where the state of a provider may be disposed when it shouldn't be disposed.